# Felix/active series encoding benchmark