--- a/pkg/querier/cardinality_analysis_handler.go
+++ b/pkg/querier/cardinality_analysis_handler.go
@@ -9,8 +9,9 @@ import (
 
 	"github.com/grafana/dskit/httpgrpc"
 	"github.com/grafana/dskit/tenant"
+	jsoniter "github.com/json-iterator/go"
 	"github.com/pkg/errors"
-	"github.com/prometheus/prometheus/model/labels"
+	"github.com/prometheus/prometheus/web/api/v1"
 
 	"github.com/grafana/mimir/pkg/cardinality"
 	ingester_client "github.com/grafana/mimir/pkg/ingester/client"
@@ -106,7 +107,16 @@ func ActiveSeriesCardinalityHandler(distributor Distributor, limits *validation.
 			return
 		}
 
-		util.WriteJSONResponse(w, activeSeriesResponse{res})
+		var json = jsoniter.ConfigCompatibleWithStandardLibrary
+		bytes, err := json.Marshal(v1.Response{Data: res})
+		if err != nil {
+			http.Error(w, err.Error(), http.StatusInternalServerError)
+		}
+
+		w.Header().Set("Content-Type", "application/json")
+
+		// Nothing we can do about this error, so ignore it.
+		_, _ = w.Write(bytes)
 	})
 }
 
@@ -236,8 +246,4 @@ type labelNamesCardinality struct {
 type labelValuesCardinalityResponse struct {
 	SeriesCountTotal uint64                  `json:"series_count_total"`
 	Labels           []labelNamesCardinality `json:"labels"`
-}
-
-type activeSeriesResponse struct {
-	Data []labels.Labels `json:"data"`
 }